Moved non-base units config
